fix: Change disqus code to use identifier based on wordpress_id and wordpress_guid.
This change corrects an issue where the disqus code was not setting the
proper identifier used by the previous site.  This correcting entailed
modifiying all posts to change the generated 'id' and 'guid' front
matter variables to 'wordpress_id' and 'wordpress_guid' respectively.
This was necessary because Jekyll overwrites the 'id' variable with a
page slug.